Tests/LibWeb/Text/expected/Editing/backspace-cancelled-keydown.txt

editor text: a
events: keydown:Backspace, beforeinput:deleteContentBackward
